Telecom Deployment
Telecom Billing System is based on a concept to generate telecom office record, to add, update and maintain their daily records. From this system user can Add, List, Modify, Delete record as well as this program contains Payment system too which helps for the bill payments of the customers. All the records can be managed and payments can be done with in a short period of time. This whole project is designed in ‘C’ Programming language & different variables and strings have been used for the development of this project. It’s easy to operate and understand by users.

Features :

Login System with Proper Validations

Easy Payments through the system

Systematic arrangement of records

Easy to add, modify, remove, list records